+++ Hash/sha/sha1.h
@@ -1,8 +1,8 @@
 #ifndef _SHA1_H
 #define _SHA1_H
 
-#include "../../common/common.h"
 
+#include <common/common.h>
 #define BLOCK_SIZE 64
 #define DIGEST_SIZE 20
 
@@ -14,7 +14,7 @@ typedef struct _HashState {
 } HashState;
 
 ErrCrypto SHA1_init(HashState* pHashState);
-ErrCrypto SHA1_update(HashState* pHashState, const uint64_t* pBuf, uint64_t nLen);
+ErrCrypto SHA1_update(HashState* pHashState, const uint8_t* pBuf, uint64_t nLen);
 ErrCrypto SHA1_digest(HashState* pHashState, uint8_t* digest, int nDigest/* DIGEST_SIZE */);
 void test_sha1();
 #endif
